% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_models.R
\name{anaylse_trials}
\alias{anaylse_trials}
\title{Analyze microbiome trials using GAMLSS}
\usage{
anaylse_trials(
  trials,
  keyterms,
  adjterms,
  formula,
  model_type = "BEZI",
  cutoff = 7,
  prevalence_threshold = 5,
  new_si = NULL,
  level = "Genus"
)
}
\arguments{
\item{trials}{A list of phyloseqSparse objects representing trials}

\item{keyterms}{Key terms for model formula}

\item{adjterms}{Adjustment terms for model formula}

\item{formula}{Formula as string}

\item{model_type}{Model type, options are "BEZI", "ZIBE", etc.}

\item{cutoff}{Minimum sample size cutoff, trials below cutoff return NA}

\item{new_si}{Optionally provide new sample information data frame}
}
\value{
A list of data frames with model results for each trial
}
\description{
Analyze microbiome trials using GAMLSS
}
\examples{
trials <- list(trial1_ps, trial2_ps)
results <- anaylse_trials(trials, keyterms = "GroupName",
                          adjterms = "~ Age + BMI",
                          formula = "Abundance ~ GroupName",
                          model_type = "BEZI")

}
