% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_accessors.R
\name{create_id}
\alias{create_id}
\title{Create Identifier Column}
\usage{
create_id(tx, level)
}
\arguments{
\item{tx}{Data frame}

\item{level}{Column name to concatenate up to}
}
\value{
The original data frame with the new identifier column added
}
\description{
Creates a character column that concatenates selected columns with underscores.
Useful for creating identifiers.
}
\examples{
mydf <- data.frame(Col1 = letters[1:3],
                   Col2 = 1:3,
                   Col3 = c("Group1", "Group2", "Group1"))
newdf <- create_id(mydf, "Col3")

}
