% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_accessors.R
\name{extract_phyloseq}
\alias{extract_phyloseq}
\title{Filter and extract data from phyloseq}
\usage{
extract_phyloseq(physeq, TSS = TRUE, prevalence_threshold = 5, new_si = NULL)
}
\arguments{
\item{physeq}{A phyloseq object}

\item{prevalence_threshold}{Prevalence threshold (proportion of samples) for filtering OTUs}
}
\value{
A named list with 3 elements:
\itemize{
\item{\code{otu_table}}{OTU table as tibble with OTU_ID column}
\item{\code{tax_table}}{Taxonomy table as tibble with Taxonomy_ID column}
\item{\code{sample_data}}{Sample data as tibble}
}
}
\description{
Filter and extract data from phyloseq
}
\examples{
filtered_data <- extract_phyloseq(physeq, prevalence_threshold = 0.05, abundance_threshold = 0.001)

}
