% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fit_gamlsss_BEZI}
\alias{fit_gamlsss_BEZI}
\title{Fit BEZI GAMLSS Model}
\usage{
fit_gamlsss_BEZI(X, y, the_formula, key_term, robust = FALSE)
}
\arguments{
\item{X}{Data frame containing model matrix}

\item{y}{Response variable name}

\item{the_formula}{Formula for gamlss model}

\item{key_term}{Name of key variable of interest}

\item{robust}{Logical, use robust standard errors}
}
\value{
Tibble with model summary statistics if model converges, NA otherwise
}
\description{
Fits a Beta Zero-Inflated GAMLSS model using the BEZI distribution.
Performs stepwise model selection and extracts key model summary statistics.
}
