% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fit_gamlsss_ZINBI}
\alias{fit_gamlsss_ZINBI}
\title{Fit Zero-Inflated Negative Binomial GAMLSS Model}
\usage{
fit_gamlsss_ZINBI(X, y, the_formula, key_term, robust = FALSE)
}
\arguments{
\item{X}{Data frame containing model matrix}

\item{y}{Response variable name}

\item{the_formula}{Formula for gamlss model}

\item{key_term}{Name of key variable of interest}

\item{robust}{Logical, use robust standard errors}
}
\description{
Fit Zero-Inflated Negative Binomial GAMLSS Model
}
