% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{gamlsss_by_sv}
\alias{gamlsss_by_sv}
\title{Fit GAMLSS Model By Species}
\usage{
gamlsss_by_sv(svs, X, model_params, model_type = "BEZI")
}
\arguments{
\item{svs}{Species identifiers}

\item{X}{Model data frame}

\item{model_params}{Model dictionary}

\item{model_type}{Model type}
}
\value{
Data frame with model results for each species
}
\description{
Fits GAMLSS to each species in a phylogenetic sequence.
}
