% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{model_dict}
\alias{model_dict}
\title{Create Model Dictionary}
\usage{
model_dict(keyterms, formula, data, adjterms = NULL)
}
\arguments{
\item{keyterms}{Key term}

\item{formula}{Formula string}

\item{data}{Model data frame}

\item{adjterms}{Adjustment terms}
}
\value{
List containing model terms
}
\description{
Validates model terms and creates model dictionary.
}
