% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_anaylsis.R
\name{run_bayesmeta}
\alias{run_bayesmeta}
\title{Run Bayesian Meta-Analysis for Different Taxonomic Levels}
\usage{
run_bayesmeta(results)
}
\arguments{
\item{results}{A list of data frames, each representing results from a different study.}
}
\value{
A tibble with Bayesian meta-analysis results for each taxonomic identifier.
}
\description{
This function aggregates results from different studies and runs a Bayesian meta-analysis for each unique taxonomic identifier.
}
\examples{
# Example usage
# run_bayesmeta(results = list_of_study_results)
}
