% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_anaylsis.R
\name{run_metareg}
\alias{run_metareg}
\title{Run Meta-Regression for Different Taxonomic Levels}
\usage{
run_metareg(results, sm = "OR")
}
\arguments{
\item{results}{A list of data frames, each representing results from a different study.}

\item{sm}{A character string specifying the summary measure; default is 'OR' (odds ratio).}
}
\value{
A tibble with meta-regression results for each taxonomic identifier.
}
\description{
This function aggregates results from different studies and runs meta-regression for each unique taxonomic identifier.
}
\examples{
# Example usage
# run_metareg(results = list_of_study_results, sm = 'RR')
}
